Combobox control

+++ fx-support/src/main/scala/com/alphasystem/fx/ui/util/FontUtilities.scala
@@ -15,16 +15,16 @@ object FontUtilities {
   private val ArabicRegularFontSizeKey = "arabic.regular.font.size"
   private val ArabicHeadingFontSizeKey = "arabic.heading.font.size"
   private val EnglishRegularFontSizeKey = "english.regular.font.size"
-  private val DefaultArabicFontSize = 20L
-  private val DefaultArabicHeadingFontSize = 40L
-  private val DefaultEnglishFontSize = 12L
+  private val DefaultArabicFontSize = 20.0
+  private val DefaultArabicHeadingFontSize = 40.0
+  private val DefaultEnglishFontSize = 12.0
 
   private lazy val families = Font.getFamilies.asScala.toList
   lazy val ArabicFontName: String = getDefaultArabicFontName(families)
   lazy val EnglishFontName: String = getDefaultEnglishFont(families)
-  lazy val ArabicRegularFontSize: Long = getArabicRegularFontSize
-  lazy val ArabicHeadingFontSize: Long = getArabicHeadingFontSize
-  lazy val EnglishFontSize: Long = getEnglishRegularFontSize
+  lazy val ArabicRegularFontSize: Double = getArabicRegularFontSize
+  lazy val ArabicHeadingFontSize: Double = getArabicHeadingFontSize
+  lazy val EnglishFontSize: Double = getEnglishRegularFontSize
 
   private def getDefaultArabicFontName(families: List[String]): String = {
     val fontName = System.getProperty(ArabicFontNameKey)
@@ -38,9 +38,9 @@ object FontUtilities {
   private def getDefaultEnglishFont(families: List[String]) =
     if families.contains(Candara) then Candara else Arial
 
-  private def getFontSize(propertyName: String, defaultSize: Long) = {
+  private def getFontSize(propertyName: String, defaultSize: Double) = {
     val defaultValue = String.valueOf(defaultSize)
-    System.getProperty(propertyName, defaultValue).toLong
+    System.getProperty(propertyName, defaultValue).toDouble
   }
 
   private def getArabicRegularFontSize =
